+++ src/strigui/input.clj
@@ -1,6 +1,8 @@
 (ns strigui.input
   (:require [strigui.widget :as wdg]
-            [strigui.box :as b]))
+            [strigui.box :as b])
+  (:import [java.awt.datatransfer Clipboard StringSelection DataFlavor]
+           [java.awt Toolkit]))
 
 (defrecord Input [name value props]
   wdg/Widget
@@ -52,13 +54,33 @@
     (let [input (->Input name text props)]
       input))
 
+(defn copy->clipboard!
+  "Copies the value of the current input widget to the systems clipboard"
+  [widget]
+  (let [clip-board ^Clipboard (.getSystemClipboard (Toolkit/getDefaultToolkit))
+        selection (StringSelection. ^String (:value widget))]
+    (.setContents clip-board selection selection) 
+    widget))
+
+(defn clipboard->widget!
+  "Paste the clipboard contents to the widgets value"
+  [widget]
+  (let [clip-board ^Clipboard (.getSystemClipboard (Toolkit/getDefaultToolkit))
+        content ^String (.toString (.getTransferData (.getContents clip-board nil) DataFlavor/stringFlavor))]
+    (update widget :value #(str % content))))
+
 (defn handle-key-pressed
-  [widget char code]
+  [widget char code prev-code] 
+  (if (= prev-code 17)
+    (case code 
+      67 (copy->clipboard! widget)
+      86 (clipboard->widget! widget)
+      widget)
     (let [box-with-new-input (key-pressed widget char code)]
           ;box-with-new-input (assoc-in box-with-new-input [:props :selected?] (or (not= code 10) ;;enter
           ;                                                                       (not= code 9)))] ;; tab
-      box-with-new-input))
+      box-with-new-input)))
 
 (defmethod wdg/widget-event [strigui.input.Input :key-pressed]
-  [_ canvas widgets widget char code]
-  (assoc widgets (:name widget) (handle-key-pressed widget char code)))
+  [_ canvas widgets widget char code prev-code]
+  (assoc widgets (:name widget) (handle-key-pressed widget char code prev-code)))