+++ CHANGES.md
@@ -12,6 +12,7 @@
 * give option for widget highlights to be a combination of :border :none :alpha (:alpha with add a transparent layer 
 in the highlight color on top)
 * highlight border size can be adjusted via key :highlight-border-size accepting a non negative integer value
+* copy and paste via ctrl+c and ctrl+v for input fields
 * TODO: add optional header and support multi columns for list widgets
 * TODO: better documentation